server: changed validators structure
