+++ server/src/routes.ts
@@ -2,38 +2,38 @@ import { Router } from 'express';
 
 import authMiddleware from './app/middlewares/auth';
 
+// Controllers
 import SessionController from './app/controllers/SessionController';
 import UserController from './app/controllers/UserController';
 import RecipientController from './app/controllers/RecipientsController';
 
-import userStoreValidation from './app/validators/userStore';
-import sessionStoreValidation from './app/validators/sessionStore';
-import recipientStoreValidation from './app/validators/recipientStore';
-import recipientUpdateValidation from './app/validators/recipientUpdate';
-import recipientShowAndDestroyValidation from './app/validators/recipientShowAndDestroy';
+// Input validators
+import RecipientValidator from './app/validators/RecipientValidator';
+import SessionValidator from './app/validators/SessionValidator';
+import UserValidator from './app/validators/UserValidator';
 
 const routes = Router();
 
-routes.post('/sessions', sessionStoreValidation, SessionController.store);
+routes.post('/sessions', SessionValidator.store, SessionController.store);
 
 routes.use(authMiddleware);
-routes.post('/users', userStoreValidation, UserController.store);
+routes.post('/users', UserValidator.store, UserController.store);
 
 routes.get('/recipients', RecipientController.index);
 routes.get(
   '/recipients/:recipient_id',
-  recipientShowAndDestroyValidation,
+  RecipientValidator.show,
   RecipientController.show
 );
-routes.post('/recipients', recipientStoreValidation, RecipientController.store);
+routes.post('/recipients', RecipientValidator.store, RecipientController.store);
 routes.put(
   '/recipients/:recipient_id',
-  recipientUpdateValidation,
+  RecipientValidator.update,
   RecipientController.update
 );
 routes.delete(
   '/recipients/:recipient_id',
-  recipientShowAndDestroyValidation,
+  RecipientValidator.destroy,
   RecipientController.destroy
 );
